requirements.txt
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.2.post1
pep8==1.7.1
pycodestyle==2.5.0
pyflakes==2.1.1
Werkzeug==0.15.1
